ResponseStatus alterado para alguns metodos do controller

+++ src/main/java/org/javaexercises/listaprodutosapp/Controller/ProdutoController.java
@@ -5,6 +5,7 @@ import org.javaexercises.listaprodutosapp.Exceptions.ProdutoNotFoundExecption;
 import org.javaexercises.listaprodutosapp.Repository.ProdutoRepository;
 import org.javaexercises.listaprodutosapp.model.Produto;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -26,13 +27,13 @@ public class ProdutoController {
     }
 
     @GetMapping("/{id}")
-    Produto findById(@PathVariable int id) throws ProdutoNotFoundExecption {
+    ResponseEntity<Produto> findById(@PathVariable int id) throws ProdutoNotFoundExecption {
         Optional<Produto> produto = produtoRepository.findById(id);
         if(produto.isEmpty()){
             throw new ProdutoNotFoundExecption();
         }
 
-        return produto.get();
+        return ResponseEntity.ok(produto.get());
     }
 
     @ResponseStatus(HttpStatus.CREATED)
@@ -43,13 +44,21 @@ public class ProdutoController {
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @PutMapping("/{id}")
-    void update(@Valid @RequestBody Produto produto, @PathVariable Integer id){
+    void update(@Valid @RequestBody Produto produto, @PathVariable Integer id) throws ProdutoNotFoundExecption {
+        Optional<Produto> produtoDatabase = produtoRepository.findById(id);
+        if(produtoDatabase.isEmpty()){
+            throw new ProdutoNotFoundExecption();
+        }
         produtoRepository.save(produto);
     }
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/{id}")
-    void delete(@PathVariable Integer id) {
+    void delete(@PathVariable Integer id) throws ProdutoNotFoundExecption {
+        Optional<Produto> produtoDatabase = produtoRepository.findById(id);
+        if(produtoDatabase.isEmpty()){
+            throw new ProdutoNotFoundExecption();
+        }
         produtoRepository.delete(produtoRepository.findById(id).get());
     }
 }
